#62 make LightingPoint standalone class
